+++ platform-sdk/consensus-utility/src/main/java/org/hiero/consensus/io/NoOpRecycleBin.java
@@ -1,10 +1,9 @@
 // SPDX-License-Identifier: Apache-2.0
-package com.swirlds.common.io.utility;
+package org.hiero.consensus.io;
 
 import edu.umd.cs.findbugs.annotations.NonNull;
 import java.io.IOException;
 import java.nio.file.Path;
-import org.hiero.consensus.io.RecycleBin;
 
 /**
  * A no-op {@link RecycleBin} implementation.
