+++ platform-sdk/consensus-utility/src/testFixtures/java/module-info.java
@@ -2,11 +2,12 @@
 open module org.hiero.consensus.utility.test.fixtures {
     exports org.hiero.consensus.test.fixtures;
     exports org.hiero.consensus.test.fixtures.crypto;
+    exports org.hiero.consensus.test.fixtures.io;
 
     requires transitive com.hedera.pbj.runtime;
     requires transitive org.hiero.base.crypto;
+    requires transitive org.hiero.consensus.utility;
     requires org.hiero.base.utility.test.fixtures;
     requires org.hiero.consensus.model;
-    requires org.hiero.consensus.utility;
     requires static transitive com.github.spotbugs.annotations;
 }
