+++ platform-sdk/consensus-utility/src/testFixtures/java/org/hiero/consensus/test/fixtures/io/TestRecycleBin.java
@@ -1,5 +1,5 @@
 // SPDX-License-Identifier: Apache-2.0
-package com.swirlds.common.test.fixtures;
+package org.hiero.consensus.test.fixtures.io;
 
 import edu.umd.cs.findbugs.annotations.NonNull;
 import java.io.IOException;
