+++ platform-sdk/swirlds-common/src/main/java/module-info.java
@@ -6,7 +6,6 @@ module com.swirlds.common {
     exports com.swirlds.common.context;
     exports com.swirlds.common.io.config;
     exports com.swirlds.common.io.exceptions;
-    exports com.swirlds.common.io.utility;
     exports com.swirlds.common.merkle.synchronization;
     exports com.swirlds.common.merkle.synchronization.streams;
     exports com.swirlds.common.merkle.synchronization.task;
@@ -19,8 +18,6 @@ module com.swirlds.common {
     exports com.swirlds.common.merkle.synchronization.stats;
 
     opens com.swirlds.common.utility to
-            com.fasterxml.jackson.databind;
-    opens com.swirlds.common.io.utility to
             com.fasterxml.jackson.databind;
 
     requires transitive com.swirlds.base;
