+++ platform-sdk/swirlds-platform-core/src/test/java/com/swirlds/platform/event/EventMigrationTest.java
@@ -1,7 +1,6 @@
 // SPDX-License-Identifier: Apache-2.0
 package com.swirlds.platform.event;
 
-import com.swirlds.common.io.utility.NoOpRecycleBin;
 import com.swirlds.config.extensions.test.fixtures.TestConfigBuilder;
 import com.swirlds.platform.recovery.internal.EventStreamSingleFileIterator;
 import edu.umd.cs.findbugs.annotations.NonNull;
@@ -21,6 +20,7 @@ import org.hiero.base.crypto.Hash;
 import org.hiero.consensus.constructable.ConstructableRegistration;
 import org.hiero.consensus.crypto.DefaultEventHasher;
 import org.hiero.consensus.io.IOIterator;
+import org.hiero.consensus.io.NoOpRecycleBin;
 import org.hiero.consensus.model.event.EventDescriptorWrapper;
 import org.hiero.consensus.model.event.PlatformEvent;
 import org.hiero.consensus.pces.config.PcesConfig_;
